FIX: SessionDelegator 삭제, WCSessionDelegate을 AppDelegate, WKExtensionDelegator가 준수
