+++ Dobby/DobbyMini WatchKit Extension/WKExtensionDelegator.swift
@@ -5,30 +5,37 @@
 //  Created by yongmin lee on 1/12/23.
 //
 
-
 import WatchKit
 import WatchConnectivity
 
 class WKExtensionDelegator: NSObject, WKExtensionDelegate {
-    
-    let sessionDelegator = SessionDelegator(
-        localStorage: LocalStorageServiceImpl.shared
-    )
+    let localStorage = LocalStorageServiceImpl.shared
     
     func applicationDidFinishLaunching() {
-        print("debug : WKExtensionDelegator applicationDidFinishLaunching " )
-        WCSession.default.delegate = self.sessionDelegator
+        WCSession.default.delegate = self
         WCSession.default.activate()
     }
     
-    
     func applicationDidBecomeActive() {
-        print("debug : watch applicationDidBecomeActive @@ 21" )
         let receiveData = WCSession.default.receivedApplicationContext
         if receiveData.isEmpty == false {
             
-        } else {
-            
         }
+    }
+}
+
+extension WKExtensionDelegator: WCSessionDelegate {
+    func session(
+        _ session: WCSession,
+        activationDidCompleteWith activationState: WCSessionActivationState,
+        error: Error?
+    ) {
+    }
+    
+    func session(
+        _ session: WCSession,
+        didReceiveApplicationContext applicationContext: [String : Any]
+    ) {
+        //
     }
 }
